requirements.txt
dnspython
requests
python-dotenv